+++ tests/testthat/test-istr.R
@@ -1,13 +1,12 @@
-test_that("outputs the expected columns", {
+test_that("outputs common output columns", {
   companies <- slice(istr_companies, 1)
   scenario <- slice(istr_weo_2022, 1)
   ep_weo <- slice(istr_ep_weo, 1)
 
   out <- istr(companies, scenario, ep_weo)
 
-  expect_true(all(common_output_columns() %in% names(out)))
-  expect_true(any(grepl("score", names(out))))
-  expect_equal(names(out)[1:3], c("id", "transition_risk", "score_aggregated"))
+  expected <- common_output_columns()
+  expect_equal(names(out)[1:4], expected)
 })
 
 test_that("outputs the expected columns", {